azure-iot-main/edgesol/edgesol/modules/PythonModule/requirements.txt
azure-iot-device~=2.3.0